Fix OSX python buildout library detection.
Patch by Jonas Baumann:

1) Support for [tests] extras:
We usually use the extras "tests" (plural) for declaring dependencies, but the
dependencychecker expects "test" (singular). I propose to support both.
The best way would be to support tests_requires but it seems not to be written
to the egg-infos.

2) Extended standard library detection:
I have python installed on my OS X using the collective/buildout.python [1] which
is a common way to install python in the plone world at least.

This installation seems to put standard library modules into two different directories:
    >>> import os
    >>> os
    <module 'os' from '/Users/jone/projects/python/buildout.python/python-2.4/lib/python2.4/os.pyc'>
    >>> import urllib
    >>> urllib
    <module 'urllib' from '/Users/jone/projects/python/buildout.python/parts/opt/lib/python2.4/urllib.pyc'>

So with this python installed it does not detect things like xml, urllib or StringIO
as standard library modules. So I've added an additional module detection based
on urllib which solves the problem.