[DottedName] dunder methods: membership
Add __contains__ method to DottedName class so that it can be used in membership testing,
i.e. if an object is part of another one, or in DottedName's use case, if a DottedName is part of another
(plone.app.dexterity.utils being part of plone.app.dexterity).

This will make DottedName objects more pythonic.