[DottedName] Fix pypi names
There are quite a few packages on pypi that have a prefix that is not part of the package namespace,
i.e. python-dateutil (that when imported is just dateutil).

Another common difference is dashes that are turned into underscores.

Only do this changes on requirements,
plain imports found by our extractors in modules.py can not have these.