feat: umbrella distributions (the actual fix)
When checking if an import, i.e. `zope.interface.Interface` and/or
a requirement, i.e. `Products.CMFCore` is one contained by the other,
always return `False` if one of them is `Zope`.

`Zope` is an umbrella distribution, i.e. a distribution that does not
provide packages that are part of its name, as `zope.container` could
be.

That means that any umbrella distribution can only be used via a user
mapping.