feat: handle modern distributions
Distributions that opted to define their configuration via `setup.cfg`
or `pyproject.toml` would not work at all with `z3c.dependencychecker`.

This commit solves it, by checking if either of the 3 files are present:
the legacy, but widely used `setup.py`, the mildly modern `setup.cfg` or
the new `pyproject.toml`.