feat: support implicit namespaces
Since Python 3.3 one does not need to create `__init__.py` files to mark
a folder to be part of a namespace, but the simple presence of a python
file deep inside it is enough to create it.

Moreover, since `setuptools` 62.3.0 it is officially discouraged to use
the old `pkg_resources` namespace technique.